# App/Kit_ListingMS/build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Size": {
				"type": "string",
				"enum": [
					"xsmall",
					"small",
					"medium",
					"large",
					"xlarge"
				]
			},
			"KitListing": {
				"properties": {
					"seller": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"size": {
						"$ref": "#/components/schemas/Size"
					},
					"colors": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"image": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"listed": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"seller",
					"title",
					"description",
					"size",
					"colors",
					"price",
					"id",
					"listed"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NewSellerListing": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"size": {
						"$ref": "#/components/schemas/Size"
					},
					"colors": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"image": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"description",
					"size",
					"colors",
					"price"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ListingPatch": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"size": {
						"$ref": "#/components/schemas/Size"
					},
					"colors": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"image": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"KeyCreated": {
				"properties": {
					"id": {
						"type": "string"
					},
					"prefix": {
						"type": "string"
					},
					"plaintext": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"created_at": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"prefix",
					"plaintext",
					"label",
					"created_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateKeyRequest": {
				"properties": {
					"label": {
						"type": "string"
					}
				},
				"required": [
					"label"
				],
				"type": "object",
				"additionalProperties": false
			},
			"KeyMetadata": {
				"properties": {
					"id": {
						"type": "string"
					},
					"prefix": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"created_at": {
						"type": "string"
					},
					"revoked_at": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"prefix"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NewKitListing": {
				"properties": {
					"seller": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"size": {
						"$ref": "#/components/schemas/Size"
					},
					"colors": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"image": {
						"type": "string"
					}
				},
				"required": [
					"seller",
					"title",
					"description",
					"size",
					"colors",
					"price"
				],
				"type": "object",
				"additionalProperties": false
			},
			"KitListingPatch": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"size": {
						"$ref": "#/components/schemas/Size"
					},
					"colors": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"image": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"apiKey": {
				"type": "http",
				"scheme": "bearer"
			},
			"jwe": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"info": {
		"title": "kit_listing-ms",
		"version": "0.0.0",
		"contact": {}
	},
	"paths": {
		"/seller/listings": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/KitListing"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorised"
					}
				},
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NewSellerListing"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/KitListing"
									},
									"type": "array"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorised"
					}
				},
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/seller/listings/{id}": {
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/KitListing"
								}
							}
						}
					},
					"400": {
						"description": "Invalid ID format"
					},
					"401": {
						"description": "Unauthorised"
					},
					"404": {
						"description": "Not found"
					}
				},
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ListingPatch"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Remove",
				"responses": {
					"204": {
						"description": "No Content"
					},
					"400": {
						"description": "Invalid ID format"
					},
					"401": {
						"description": "Unauthorised"
					},
					"404": {
						"description": "Not found"
					}
				},
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/seller/keys": {
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/KeyCreated"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorised"
					}
				},
				"security": [
					{
						"jwe": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateKeyRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/KeyMetadata"
									},
									"type": "array"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorised"
					}
				},
				"security": [
					{
						"jwe": []
					}
				],
				"parameters": []
			}
		},
		"/seller/keys/{id}": {
			"delete": {
				"operationId": "Revoke",
				"responses": {
					"204": {
						"description": "No Content"
					},
					"400": {
						"description": "Invalid ID format"
					},
					"401": {
						"description": "Unauthorised"
					},
					"404": {
						"description": "Not found"
					}
				},
				"security": [
					{
						"jwe": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/kit-listing": {
			"get": {
				"operationId": "GetAllKitListings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/KitListing"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sellerId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateNewKitListing",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/KitListing"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorised"
					}
				},
				"security": [
					{
						"jwe": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NewKitListing"
							}
						}
					}
				}
			}
		},
		"/kit-listing/{id}": {
			"get": {
				"operationId": "GetKitListingById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/KitListing"
								}
							}
						}
					},
					"400": {
						"description": "Invalid ID format"
					},
					"404": {
						"description": "Not found"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "EditKitListing",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/KitListing"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwe": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/KitListingPatch"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/api/v0"
		}
	]
}